Rename renderers/latex|typst|markdown → serializers/
Serializers convert YAML data to text strings — they don't render anything.
Rendering (PDF, HTML) happens in a separate external tool (Pandoc WASM, Typst WASM, Docsify).
Preact stays in renderers/ — it's the only one that actually outputs to the DOM.
Added serializers/README.md explaining the distinction.
