+++ packages/docsify-plugin/src/renderers/index.js
@@ -1,14 +1,1 @@
-export { latexRenderers } from './latex.js';
-export { typstRenderers } from './typst.js';
-export { markdownRenderers } from './markdown.js';
 export { preactRenderer, buildTransforms } from './preact.js';
-
-import { latexRenderers } from './latex.js';
-import { typstRenderers } from './typst.js';
-import { markdownRenderers } from './markdown.js';
-
-export const exportFormats = {
-  latex: latexRenderers,
-  typst: typstRenderers,
-  markdown: markdownRenderers,
-};
