+++ packages/docsify-plugin/src/serializers/latex.js
